feat: add new programming language pages and integrate socket.io for real-time updates
- Added C# page with code validation and victory handling.
- Added C++ page with code validation and victory handling.
- Added Flutter page with code validation and victory handling.
- Added JSX page with code validation and victory handling.
- Added PHP page with code validation and victory handling.
- Integrated socket.io for real-time communication on victory events.
- Updated package.json and package-lock.json to include axios dependency.
