+++ package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "axios": "^1.9.0",
     "next": "^15.2.4",
     "postcss": "^8.5.3",
     "react": "^19.0.0",
